# specs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GameBot API",
    "description": "Canonical GameBot endpoints under /api",
    "version": "v1"
  },
  "paths": {
...
    "/api/sessions": {
      "post": {
        "tags": [
          "Sessions",
          "Sessions"
        ],
        "summary": "Create a session",
        "operationId": "CreateSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              },
              "example": {
                "gameId": "game-123",
                "adbSerial": "emulator-5554"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                },
                "example": {
                  "id": "session-123",
                  "status": "PENDING",
                  "gameId": "game-123"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "tags": [
          "Sessions",
          "Sessions"
        ],
        "summary": "Get a session",
        "operationId": "GetSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailSchema"
                },
                "example": {
                  "id": "session-123",
                  "status": "RUNNING",
                  "uptime": 42,
                  "health": "HEALTHY",
                  "gameId": "game-123"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sessions",
          "Sessions"
        ],
        "summary": "Stop a session",
        "operationId": "StopSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStopSchema"
                },
                "example": {
                  "stopped": true
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/start": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Start a session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              },
              "example": {
                "primitiveAction": {
                  "type": "connect-to-game",
                  "schemaVersion": "v1",
                  "payload": {
                    "gameId": "game-123",
                    "adbSerial": "emulator-5554"
                  }
                }
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StartSessionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSessionResponse"
                },
                "example": {
                  "sessionId": "session-abc",
                  "runningSessions": [
                    {
                      "sessionId": "session-abc",
                      "gameId": "game-123",
                      "emulatorId": "emulator-5554",
                      "startedAtUtc": "2026-01-28T10:00:00Z",
                      "lastHeartbeatUtc": "2026-01-28T10:02:00Z",
                      "status": "running"
                    },
                    {
                      "sessionId": "session-def",
                      "gameId": "game-456",
                      "emulatorId": "emulator-9999",
                      "startedAtUtc": "2026-01-28T09:58:00Z",
                      "lastHeartbeatUtc": "2026-01-28T10:01:30Z",
                      "status": "running"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/stop": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Stop a session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopSessionRequest"
              },
              "example": {
                "sessionId": "session-abc"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StopSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StopSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopSessionResponse"
                },
                "example": {
                  "stopped": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string",
            "nullable": true
          },
          "gamePath": {
            "type": "string",
            "nullable": true
          },
          "adbSerial": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateSessionResponse": {
        "required": [
          "gameId",
          "id",
          "sessionId",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "gameId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PrimitiveActionRequest": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "schemaVersion": {
            "type": "string",
            "nullable": true
          },
          "payload": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RunningSessionDto": {
        "required": [
          "emulatorId",
          "gameId",
          "sessionId"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "gameId": {
            "type": "string",
            "nullable": true
          },
          "emulatorId": {
            "type": "string",
            "nullable": true
          },
          "startedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "lastHeartbeatUtc": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/RunningSessionStatus"
          },
          "captureRateFps": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RunningSessionStatus": {
        "enum": [
          "Running",
          "Stopping"
        ],
        "type": "string"
      },
...
      "SessionDetailSchema": {
        "required": [
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "uptime": {
            "type": "integer",
            "format": "int64"
          },
          "health": {
            "type": "string",
            "nullable": true
          },
          "gameId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SessionStopSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StartSessionRequest": {
        "required": [
          "primitiveAction"
        ],
        "type": "object",
        "properties": {
          "primitiveAction": {
            "$ref": "#/components/schemas/PrimitiveActionRequest"
          }
        },
        "additionalProperties": false
      },
      "StartSessionResponse": {
        "required": [
          "runningSessions",
          "sessionId"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "runningSessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunningSessionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StopSessionRequest": {
        "required": [
          "sessionId"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StopSessionResponse": {
        "type": "object",
        "properties": {
          "stopped": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}